perf(cli): load semver lazily for update checks
