+++ packages/cli/src/services/updater.ts
@@ -5,12 +5,10 @@ import { Context, Duration, Effect, FileSystem, Layer } from "effect"
 import { ChildProcess } from "effect/unstable/process"
 import { parse, type ParseError } from "jsonc-parser"
 import path from "node:path"
-import semver from "semver"
 
 declare const OPENCODE_CLI_NAME: string | undefined
 
 export type Policy = boolean | "notify"
-export type Action = "none" | "upgrade"
 type Method = "npm" | "pnpm" | "bun" | "yarn"
 
 const packageName =
@@ -32,14 +30,6 @@ export function decodePolicy(text: string): Policy | undefined {
   if (errors.length || typeof input !== "object" || input === null || !("autoupdate" in input)) return
   const value = input.autoupdate
   if (typeof value === "boolean" || value === "notify") return value
-}
-
-export function action(current: string, latest: string, policy: Policy): Action {
-  if (policy === false) return "none"
-  if (!semver.valid(current) || !semver.valid(latest) || semver.eq(latest, current)) return "none"
-  // Major upgrades are never installed automatically.
-  if (semver.major(latest) !== semver.major(current)) return "none"
-  return "upgrade"
 }
 
 export const layer = Layer.effect(
@@ -150,6 +140,7 @@ export const layer = Layer.effect(
             current: OPENCODE_VERSION,
             latest: version,
           })
+          const { action } = yield* Effect.promise(() => import("./updater-action"))
           const next = action(OPENCODE_VERSION, version, policy)
           if (next === "none") return yield* Effect.logInfo("update check done", { action: "up-to-date" })
           const detected = yield* method()
